[FIX] Configure DMA region for RPi3/RPi4
RPi3/RPi4 require DMA for their mailbox component. This is a platform dependent configuration, that is currently not supported in the macros.
This commit moves SdHostController_INSTANCE_CONFIGURE_BY_INDEX into the platform specific files and adds DMA configuration for RPi3/RPi4. The default value of 4096 bytes for the DMA size is used, as required by mailbox_init().